src/requirements.txt
absl-py==0.10.0
aiohttp==3.6.2
aiohttp-jinja2==1.2.0
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.2.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
Faker==4.1.3
future==0.18.2
gast==0.3.3
google-auth==1.21.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==2.0.0
Jinja2==2.11.2
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.2
multidict==4.7.6
mxnet==1.5.1
networkx==2.5
numpy==1.19.2
oauthlib==3.1.0
onnx==1.7.0
opt-einsum==3.3.0
pandas==1.1.4
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
test-generator==0.1.1
text-unidecode==1.3
torch==1.7.0
torchvision==0.8.1
tqdm==4.51.0
typing-extensions==3.7.4.3
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.5.1
yml==0.0.1
zipp==3.2.0